+++ src/components/StrategyArguments.vue
@@ -18,12 +18,17 @@
             ></b-form-input>
             </b-form-group>
 
-            <b-form-group label="Start Date" label-for="simulationBegin">
-            <b-form-datepicker id="simulationBegin" v-model="simulationBegin" :date-format-options="{ year: 'numeric', month: 'short'}" :disabled="disabled"></b-form-datepicker>
+            <b-form-group label="Benchmark" description="Symbol to compare performance against" label-for="benchmarkTickerId">
+            <b-form-input
+                id="benchmarkTickerId"
+                v-model="benchmarkTickerData"
+                type="text"
+                :disabled="disabled"
+            ></b-form-input>
             </b-form-group>
 
-            <b-form-group label="End Date" label-for="simulationEnd">
-            <b-form-datepicker id="simulationEnd" v-model="simulationEnd" :date-format-options="{ year: 'numeric', month: 'short'}" :disabled="disabled"></b-form-datepicker>
+            <b-form-group label="Date Range" label-for="simulationDates">
+              <VueCtkDateTimePicker class="datePicker" id="simulationDates" v-model="simulationDates" formatted="DD MMM YY" :custom-shortcuts="shortcuts" :disabled="disabled" :disabled-weekly="disabledDates" noWeekendDays noLabel range />
             </b-form-group>
 
             <b-button v-if="!disabled" type="submit" class="mr-2" variant="info">Submit</b-button>
@@ -33,6 +38,10 @@
 </template>
 
 <script>
+import Moment from 'moment'
+import VueCtkDateTimePicker from 'vue-ctk-date-time-picker';
+import 'vue-ctk-date-time-picker/dist/vue-ctk-date-time-picker.css';
+
 export default {
   name: 'StrategyArguments',
   props: {
@@ -40,14 +49,50 @@ export default {
     callback: Function,
     disabled: Boolean,
     begin: Date,
-    end: Date
+    end: Date,
+    benchmarkTicker: {
+      type: String,
+      default: "VFINX"
+    }
   },
   data() {
     return {
       form: {},
-      simulationBegin: new Date(1980,0,1),
-      simulationEnd: new Date(),
+      benchmarkTickerData: this.benchmarkTicker,
+      disabledDates: [0,6],
+      shortcuts: [
+        { key: 'thisYear', label: 'This year', value: 'year' },
+        { key: 'lastYear', label: 'Last year', value: '-year' },
+        { key: 'last3Years', label: 'Last 3 years', value: () => {
+          return {
+            start: Moment().subtract(3, 'years'),
+            end: Moment()
+          }}
+        },
+        { key: 'last5Years', label: 'Last 5 years', value: () => {
+          return {
+            start: Moment().subtract(5, 'years'),
+            end: Moment()
+          }}
+        },
+        { key: 'last10Years', label: 'Last 10 years', value: () => {
+          return {
+            start: Moment().subtract(10, 'years'),
+            end: Moment()
+          }}
+        },
+        { key: 'last15Years', label: 'Last 15 years', value: () => {
+          return {
+            start: Moment().subtract(15, 'years'),
+            end: Moment()
+          }}
+        }
+      ],
+      simulationDates: {"start": new Date(1980,0,1), "end": new Date()}
     }
+  },
+  components: {
+    VueCtkDateTimePicker
   },
   mounted: async function() {
       Object.entries(this.spec).forEach(elem => {
@@ -55,10 +100,10 @@ export default {
         this.form[elem.arg] = elem.inpdefault
       })
       if (this.begin != 0) {
-        this.simulationBegin = this.begin
+        this.simulationDates.start = this.begin
       }
       if (this.end != 0) {
-        this.simulationEnd = this.end
+        this.simulationDates.end = this.end
       }
   },
   watch: {
@@ -69,21 +114,21 @@ export default {
       })
     },
     begin: function(n) {
-      this.simulationBegin = n
+      this.simulationDates.start = n
     },
     end: function(n) {
-      this.simulationEnd = n
+      this.simulationDates.end = n
     }
   },
   methods: {
       onSubmit: async function(e) {
           e.preventDefault()
-          this.$emit("execute", this.form, this.simulationBegin, this.simulationEnd)
+          this.$emit("execute", this.form, this.simulationDates.start, this.simulationDates.end, this.benchmarkTickerData)
       }
   }
 }
 </script>
 
-<!-- Add "scoped" attribute to limit CSS to this component only -->
 <style scoped>
+
 </style>