+++ src/views/PortfolioCards.vue
@@ -44,12 +44,12 @@ Vue.filter("formatPercent", function (value) {
   var v;
   if (typeof value === "number") {
     v = value * 100
-    v = v.toFixed(2)
+    v = v.toFixed(1)
     return `${v}%`
   } else if (typeof value === "object") {
     if (value["Valid"]) {
       v = value["Float64"] * 100
-      v = v.toFixed(2)
+      v = v.toFixed(1)
       return `${v}%`
     }
   }