+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2021-02-14
+### Added
+- Monthly return heatmap chart on strategy performance summary tab
+- Display investment value in portfolio tab
+- Ability to download portfolio as a CSV
+- Toggle to show log scale on return chart
+- Add max draw down to summary page and option to plot draw downs on chart
+- Add tables for CAGR, portfolio metrics, and risk metrics to summary view
+- Add benchmark comparison
+
+### Changed
+- Datepicker now supports quick shortcuts and easier navigation between years
+- Adjust precision on Portfolio Cards to match what is displayed in Summary page
+
+### Fixed
+- Change Vue.js mode to use hash tags so refresh does not get 404's
+
 ## [0.1.0] - 2021-02-08
 ### Added
 - Strategy card view
@@ -13,4 +30,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Terms of Service
 - Privacy Policy
 
+[0.2.0]: https://github.com/jdfergason/pv-frontend/releases/tag/v0.2.0
 [0.1.0]: https://github.com/jdfergason/pv-frontend/releases/tag/v0.1.0